move logic and state into the view model
